+++ src/constants.js
@@ -1,6 +1,6 @@
 export const MODULE_ID = "rnk-triggerz";
 export const MODULE_TITLE = "RNK Triggerz";
-export const MODULE_VERSION = "1.0.0";
+export const MODULE_VERSION = "1.0.1";
 export const SOCKET_CHANNEL = `module.${MODULE_ID}`;
 
 export const SETTING_KEYS = Object.freeze({